hh_bot/requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.1
asyncpg==0.29.0 